[FIX] 패키지 구조수정 및 자잘한 누락 추가
