+++ src/main/java/com/soom/account_api/domain/sign/data/dto/AccountAuthInfoDto.java
@@ -1,7 +1,7 @@
 package com.soom.account_api.domain.sign.data.dto;
 
 public record AccountAuthInfoDto(
-        String id,
+        String email,
         String password
 ) {
 }
