+++ src/main/java/com/soom/account_api/domain/sign/data/dto/TeacherSignupInfoDto.java
@@ -1,8 +1,8 @@
 package com.soom.account_api.domain.sign.data.dto;
 
 public record TeacherSignupInfoDto(
-        AccountAuthInfoDto accountAuthInfoDto,
-        AccountProfileInfoDto accountProfileInfoDto,
-        TeacherInfoDto teacherInfoDto
+        AccountAuthInfoDto authInfo,
+        AccountProfileInfoDto profileInfo,
+        TeacherInfoDto teacherInfo
 ) {
 }
