+++ src/main/java/com/soom/account_api/global/repository/AccountRepository.java
@@ -1,6 +1,6 @@
-package com.soom.account_api.domain.sign.data.repository;
+package com.soom.account_api.global.repository;
 
-import com.soom.account_api.domain.sign.data.entity.AccountEntity;
+import com.soom.account_api.global.entity.AccountEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface AccountRepository extends JpaRepository<AccountEntity, Long> {
