+++ src/main/java/com/soom/account_api/global/repository/StudentRepository.java
@@ -1,6 +1,6 @@
-package com.soom.account_api.domain.sign.data.repository;
+package com.soom.account_api.global.repository;
 
-import com.soom.account_api.domain.sign.data.entity.StudentEntity;
+import com.soom.account_api.global.entity.StudentEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface StudentRepository extends JpaRepository<StudentEntity, Long> {
